Flip the constant that controls allcmds mode.
This removes the need to mark the common case, and instead marks the
uncommon case (a command that reacts to _any_ channel). NOTE: This is
a semantic change for a number of commands, which were previously open
to all channels (eg !calc). This was deemed a bug - cf Kintinue oopsie.