Remove the never-implemented require_moderator flag on HTTP endpoints.
It's much more useful to build a page that presents itself differently for
mods and non-mods (cf req->misc->is_mod) than to have some kind of generic
"kick to login". If I find I'm redirecting a lot of pages to login, then I
can consider making a unified way of spelling that.