Handle keyword attributes on commands.
This brings echocommands and function-commands a bit closer together. My
ultimate goal is to have them be identical save that echocommands get saved
to the JSON file (and obviously will never be functions).

Deprecates the hidden_command flag (which was undocumented anyway), and
require_moderator, both in favour of spelling it with keywords instead.