Replace the simple drop-down with a full command options dialog.
Lots of moving pieces, but the upshot is that channel mods can now change a
lot more options about their commands. It assumes that the response is either
an options mapping with content, or an array/string of content, but not more
deeply nested; but that's all you can create in the web UI anyway.