Refactor Twitch logins back somewhat.
The only place where ?code= should be checked is the /twitchlogin page, not
anything that ensures login. This means we no longer risk getting an async
response (a Future or function) in a context where a mapping was expected.
(Side benefit: Access token requesting is now done asynchronously.) This
hopefully will poke the bug with HTTP 500s on login.