Clean up emote_code_to_markdown uses and related oddities.
There's no more need for a global function for emotifying user text
as it was only used by quotes, so now it's inlined (as like five
lines of code!). Notably, this removes the possibility of desync
between bot_emotes and emote_code_to_markdown, removes desprintfing
of text, and enables emote modifications (as per a TODO)!