Blanking ANY command now deletes it, not just specials and triggers.
The fact that this worked for triggers was an accident of implementation,
and working for specials was deliberate (deleting a special blanks it
anyway), so this makes it consistent. My misgivings are really about loss
of user commands due to a small misconfiguration, which is better solved
in other ways anyway (you could always have wiped a slab of a command the
same way, and there's no protection for that).