Big, sweeping change - builtin_param must now ALWAYS be an array.
Use of raw strings here was always a mess, and it's only shown itself to be
more and more messy as I start trying to make MustardScript work. It's time
to lock it in as being an array of parameters. Unfortunately this will mean
some slight semantic changes in certain situations.